FInterfaces split into their own dir and some code visual changes.
