+++ src/SettingsPage/Interfaces/ViewInterface.php
@@ -1,6 +1,6 @@
 <?php declare( strict_types = 1 ); # -*- coding: utf-8 -*-
 
-namespace Adminimize\SettingsPage;
+namespace Adminimize\SettingsPage\Interfaces;
 
 interface ViewInterface {
